requirements.txt
requests==2.32.3
torch==2.6.0
transformers==4.49.0
accelerate>=0.26.0
bitsandbytes==0.45.3
httpx==0.27.2
python-dotenv==1.0.1
